Rename people-mention-controller -> people-mention
