+++ src/ggrc-client/js/components/rich-text/people-mention/people-mention.stache
