+++ src/ggrc-client/styles/components/people-autocomplete/_people-autocomplete.scss
@@ -3,7 +3,7 @@
  * Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
  */
 
-people-mention-controller {
+people-mention {
   people-autocomplete-wrapper {
     position: relative;
     display: block;
